scrapers/enrollments/requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
distro==1.4.0
idna==2.8
lxml==4.4.1
numpy==1.17.2
pandas==0.25.1
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
tabula-py==1.4.1
urllib3==1.25.6
xlrd==1.2.0
